fix: noob-friendliness + C++ polish from 3-agent stress test
README (all 3 languages):
- Client-Only Install: show full WoW path example, add "create AddOns
  if missing", add /cc verification step
- Step 1: clarify ScriptMgr class vs UnitScript class (same file)
- Step 2: add search landmarks (end of function, if time==0 block)
- Step 3: note _registeredAddonPrefixes already exists in class

C++:
- Fix implicit float-to-uint32 truncation on GetHealthPct()
- Add missing #include <unordered_set> in cs_creature_codex.cpp

+++ server/Custom/creature_codex_sniffer.cpp
@@ -117,7 +117,7 @@ static void BroadcastToNearbyPlayers(Creature* creature, SpellInfo const* spell,
         return;
 
     // Build message: TYPE|entry|spellID|schoolMask|creatureName|hp%
-    uint32 hpPct = creature->GetHealthPct();
+    uint32 hpPct = static_cast<uint32>(creature->GetHealthPct());
     std::string msg = std::string(msgType) + "|"
         + std::to_string(entry) + "|"
         + std::to_string(spellId) + "|"
